fix: change data type

+++ model/web/user_response.go
@@ -1,7 +1,7 @@
 package web
 
 type UserResponse struct {
-	Id       uint   `json:"id"`
+	Id       uint8   `json:"id"`
 	Username string `json:"username"`
 	Name     string `json:"name"`
 	Gender   string `json:"gender"`
